requirements.txt
pypdf
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
